@@ -4,23 +4,17 @@
   "private": true,
   "proxy": "http://localhost:3000",
   "dependencies": {
+    "@material-ui/core": "^4.11.0",
     "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
+    "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "apexcharts": "^3.15.6",
-    "bootstrap": "^4.4.1",
-    "concurrently": "^5.1.0",
-    "express": "^4.17.1",
-    "local-storage-fallback": "^4.1.1",
-    "react": "^16.12.0",
-    "react-anime": "^3.0.3",
-    "react-apexcharts": "^1.3.6",
-    "react-bootstrap": "^1.0.0-beta.16",
-    "react-dom": "^16.12.0",
+    "aos": "^2.3.4",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-favicon": "0.0.17",
-    "react-icons": "^3.9.0",
-    "react-responsive": "^8.0.3",
-    "react-router-dom": "^5.1.2",
+    "react-icons": "^3.11.0",
+    "react-responsive": "^8.1.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.3.1"
   },
   "scripts": {
